requirements.txt
# Core Requirements
torch
torchvision
numpy
opencv-python
Pillow

# Curriculum Learning
scikit-image
pyyaml
tqdm

# Web Interface 
flask